requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
filelock==3.13.1
Flask==3.0.3
fsspec==2024.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.0
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
sympy==1.12
torch==2.3.1+cpu @ https://download.pytorch.org/whl/cpu
torchvision==0.18.1+cpu @ https://download.pytorch.org/whl/cpu
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
